+++ src/app/loadout-builder/item-filter.ts
@@ -3,7 +3,9 @@ import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { DimItem, PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import { bucketsToCategories } from 'app/loadout/mod-utils';
 import { ItemFilter } from 'app/search/filter-types';
-import { BucketHashes } from 'data/d2/generated-enums';
+import { compareBy } from 'app/utils/comparators';
+import { getSocketsByCategoryHash } from 'app/utils/socket-utils';
+import { BucketHashes, SocketCategoryHashes } from 'data/d2/generated-enums';
 import _ from 'lodash';
 import { doEnergiesMatch } from './mod-utils';
 import {
@@ -46,7 +48,7 @@ export function filterItems(
   const lockedModMap = _.groupBy(lockedMods, (mod) => mod.plug.plugCategoryHash);
 
   for (const bucket of LockableBucketHashes) {
-    const lockedModsByPlugCategoryHash = lockedModMap[bucketsToCategories[bucket]];
+    const lockedModsForPlugCategoryHash = lockedModMap[bucketsToCategories[bucket]];
 
     if (items[bucket]) {
       // There can only be one pinned item as we hide items from the item picker once
@@ -74,10 +76,11 @@ export function filterItems(
           matchedLockedModEnergy(
             defs,
             item,
-            lockedModsByPlugCategoryHash,
+            lockedModsForPlugCategoryHash,
             upgradeSpendTier,
             lockItemEnergyType
-          )
+          ) &&
+          hasEnoughSocketsForMods(defs, item, lockedModsForPlugCategoryHash)
       );
 
       const searchFilteredItems = excludedAndModsFilteredItems.filter(searchFilter);
@@ -94,14 +97,61 @@ export function filterItems(
 function matchedLockedModEnergy(
   defs: D2ManifestDefinitions,
   item: DimItem,
-  lockedModsByPlugCategoryHash: PluggableInventoryItemDefinition[],
+  lockedMods: PluggableInventoryItemDefinition[] | undefined,
   upgradeSpendTier: UpgradeSpendTier,
   lockItemEnergyType: boolean
 ) {
-  if (!lockedModsByPlugCategoryHash) {
+  if (!lockedMods) {
     return true;
   }
-  return lockedModsByPlugCategoryHash.every((mod) =>
+  return lockedMods.every((mod) =>
     doEnergiesMatch(defs, mod, item, upgradeSpendTier, lockItemEnergyType)
   );
+}
+
+/**
+ * This ensures the item has enough sockets for the given mods by checking the plug sets for the items sockets.
+ * It does the following
+ * 1. Get a list of plugsets from the item sockets and sorts them so the shortest plugsets are used first (needed
+ *   for artificer sockets which is a subset of the bucket specific mod sockets)
+ * 2. For each locked mod, if it can go into one of the sockets, remove that socket from the list.
+ * 3. If we ever can't find a socket, we can't fit them all.
+ */
+function hasEnoughSocketsForMods(
+  defs: D2ManifestDefinitions,
+  item: DimItem,
+  lockedMods: PluggableInventoryItemDefinition[]
+) {
+  if (!lockedMods?.length) {
+    return true;
+  }
+
+  const sockets = getSocketsByCategoryHash(item.sockets!, SocketCategoryHashes.ArmorMods);
+
+  // We get the plugSets for each item because they indicate which mods can go in which socket.
+  // Artificer sockets only plug a subset of the bucket specific mods so we sort by the size
+  // of the plugItems in the plugset so we use that first if possible.
+  const plugSets = _.compact(
+    sockets.map(
+      (socket) =>
+        socket.socketDefinition.reusablePlugSetHash &&
+        // If a socket is not plugged (even with an empty socket) we consider it disabled
+        // This needs to be checked as the 30th anniversary armour has has the Artificer socket
+        // but the API considers it to be disabled.
+        socket.plugged &&
+        defs.PlugSet.get(socket.socketDefinition.reusablePlugSetHash).reusablePlugItems.map(
+          (plugItem) => plugItem.plugItemHash
+        )
+    )
+  ).sort(compareBy((plugHashes) => plugHashes.length));
+
+  for (const mod of lockedMods) {
+    const plugSetIndex = plugSets.findIndex((set) => set.includes(mod.hash));
+    if (plugSetIndex === -1) {
+      return false;
+    }
+    plugSets.splice(plugSetIndex, 1);
+  }
+
+  return true;
 }